backend/requirements.txt
# Flask and Web Framework
Flask==2.3.2
flask-cors==4.0.0

# Database - MySQL
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.3

# Stock Data APIs
yfinance>=0.2.18
finance-datareader>=0.9.00

# Data Processing (Python 3.12 호환 버전)
pandas>=2.1.0
numpy>=1.26.0

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Date and Time
python-dateutil==2.8.2
pytz>=2023.3

# Logging
colorlog==6.7.0

# Optional: For better performance
lxml>=4.9.3
html5lib==1.1
beautifulsoup4==4.12.2

# Optional: For data caching
redis==4.6.0

flask-limiter==3.5.0
